+++ example/lib/main.dart
@@ -319,7 +319,7 @@ class _FilesScreenState extends State<FilesScreen> {
       final files = await widget.uploadcareClient.files.list(
         offset: 0,
         limit: _limit,
-        stored: false,
+        stored: null,
       );
 
       setState(() {
@@ -371,20 +371,20 @@ class _FilesScreenState extends State<FilesScreen> {
                 trailing: Row(
                   mainAxisSize: MainAxisSize.min,
                   children: <Widget>[
-                    // if (/*file.isImage*/ false)
-                    //   IconButton(
-                    //     icon: Icon(Icons.face),
-                    //     onPressed: () => Navigator.push(
-                    //       context,
-                    //       MaterialPageRoute(
-                    //         fullscreenDialog: true,
-                    //         builder: (context) => FaceDetectScreen(
-                    //           uploadcareClient: widget.uploadcareClient,
-                    //           imageId: file.id,
-                    //         ),
-                    //       ),
-                    //     ),
-                    //   ),
+                    if (file.isImage)
+                      IconButton(
+                        icon: Icon(Icons.face),
+                        onPressed: () => Navigator.push(
+                          context,
+                          MaterialPageRoute(
+                            fullscreenDialog: true,
+                            builder: (context) => FaceDetectScreen(
+                              uploadcareClient: widget.uploadcareClient,
+                              imageId: file.id,
+                            ),
+                          ),
+                        ),
+                      ),
                     IconButton(
                       icon: Icon(
                         Icons.delete,
@@ -424,7 +424,7 @@ class FaceDetectScreen extends StatefulWidget {
 }
 
 class _FaceDetectScreenState extends State<FaceDetectScreen> {
-  Future<List<Rect>> _future;
+  Future<FacesEntity> _future;
   GlobalKey _key;
 
   @override
@@ -432,7 +432,7 @@ class _FaceDetectScreenState extends State<FaceDetectScreen> {
     super.initState();
 
     _key = GlobalKey();
-    _future = widget.uploadcareClient.files.detectFaces(widget.imageId);
+    _future = widget.uploadcareClient.files.getFacesEntity(widget.imageId);
   }
 
   @override
@@ -443,16 +443,18 @@ class _FaceDetectScreenState extends State<FaceDetectScreen> {
       ),
       body: FutureBuilder(
         future: _future,
-        builder: (BuildContext context, AsyncSnapshot<List<Rect>> snapshot) {
+        builder: (BuildContext context, AsyncSnapshot<FacesEntity> snapshot) {
           if (!snapshot.hasData)
             return Center(
               child: CircularProgressIndicator(),
             );
 
-          if (snapshot.data.isEmpty)
+          if (!snapshot.data.hasFaces)
             return Center(
               child: Text('No faces has been detected'),
             );
+
+          RenderBox renderBox = context.findRenderObject();
 
           return FractionallySizedBox(
             widthFactor: 1,
@@ -463,13 +465,19 @@ class _FaceDetectScreenState extends State<FaceDetectScreen> {
                   child: Image(
                     key: _key,
                     image: UploadcareImageProvider(widget.imageId),
-                    fit: BoxFit.cover,
+                    fit: BoxFit.contain,
+                    alignment: Alignment.topCenter,
                   ),
                 ),
-                ...snapshot.data.map((face) {
-                  RenderBox renderBox = context.findRenderObject();
-                  face = face.topLeft & renderBox.size;
-
+                ...snapshot.data
+                    .getRelativeFaces(
+                  Size(
+                    renderBox.size.width,
+                    renderBox.size.width /
+                        snapshot.data.originalSize.aspectRatio,
+                  ),
+                )
+                    .map((face) {
                   return Positioned(
                     top: face.top,
                     left: face.left,